refactor model runners and remove train runner mega script
