+++ happy/graph/runners/eval_runner.py
@@ -1,159 +1,275 @@
-import os
+from dataclasses import dataclass
+from typing import Optional
 
-import numpy as np
-import pandas as pd
+import torch
+import torch.nn as nn
+from torch_geometric.data import Data
 
-from happy.hdf5 import get_embeddings_file, HDF5Dataset
-from projects.placenta.graphs.processing.process_knots import process_knts
+# Attributes added to MessagePassing/nn.Module in newer PyG/PyTorch versions
+# that are absent from models saved with older versions.
+_COMPAT_ATTR_DEFAULTS = {
+    "_lazy_load_hook": None,
+    "_is_full_backward_hook": None,
+    "_forward_pre_hooks_with_kwargs": {},
+    "_forward_hooks_with_kwargs": {},
+    "_forward_hooks_always_called": {},
+    "_state_dict_pre_hooks": {},
+    "_decomposed_layers": 1,
+}
 
+def _patch_module_getattr():
+    _original = nn.Module.__getattr__
+    def _patched(self, name: str):
+        if name in _COMPAT_ATTR_DEFAULTS:
+            default = _COMPAT_ATTR_DEFAULTS[name]
+            object.__setattr__(self, name, default)
+            return default
+        return _original(self, name)
+    nn.Module.__getattr__ = _patched
+from torch_geometric.transforms import SIGN
+from torch_geometric.loader import (
+    DataLoader,
+    NeighborLoader,
+    ShaDowKHopSampler,
+)
 
-def get_and_process_raw_data(
-    project_name, organ, project_dir, run_id, graph_params, tissue_label_tsv
-):
-    if "x_min" not in graph_params:
-        graph_params["x_min"] = 0
-        graph_params["y_min"] = 0
-        graph_params["width"] = -1
-        graph_params["height"] = -1
-    hdf5_data, tissue_class = get_raw_data(
-        project_name, organ, project_dir, run_id, graph_params, tissue_label_tsv
-    )
-    hdf5_data, tissue_class = process_raw_data(
-        organ,
-        hdf5_data,
-        tissue_class,
-        graph_params["group_knts"],
-        graph_params["random_remove"],
-        graph_params["top_conf"],
-        graph_params["standardise"],
-    )
-    return hdf5_data, tissue_class
-
-
-def get_raw_data(
-    project_name,
-    organ,
-    project_dir,
-    run_id,
-    graph_params,
-    tissue_label_tsv,
-):
-    # Get training data from hdf5 files
-    hdf5_data = get_hdf5_data(
-        project_name,
-        run_id,
-        graph_params["x_min"],
-        graph_params["y_min"],
-        graph_params["width"],
-        graph_params["height"],
-    )
-    # Get ground truth manually annotated data
-    _, _, tissue_class = get_groundtruth_patch(
-        organ,
-        project_dir,
-        graph_params["x_min"],
-        graph_params["y_min"],
-        graph_params["width"],
-        graph_params["height"],
-        tissue_label_tsv,
-    )
-    return hdf5_data, tissue_class
+from happy.models.graphsaint import GraphSAINT
+from happy.models.sign import SIGN as SIGN_MLP
+from happy.organs import Organ
+from happy.graph.enums import SupervisedModelsArg
 
 
-def process_raw_data(
-    organ,
-    hdf5_data,
-    tissue_class=None,
-    group_knts=True,
-    random_remove=False,
-    top_conf=False,
-    standardise=False,
-):
-    # Covert isolated knts into syn and turn groups into a single knt point
-    if group_knts:
-        hdf5_data, tissue_class = process_knts(organ, hdf5_data, tissue_class)
-    if top_conf:
-        hdf5_data, tissue_class = confidence_filter(hdf5_data, 0.9, 1.0, tissue_class)
-    # Remove a random percentage of the data
-    if random_remove > 0.0:
-        hdf5_data, tissue_class = random_filter(hdf5_data, random_remove, tissue_class)
-    if standardise:
-        hdf5_data = standardise_cells(hdf5_data)
-    return hdf5_data, tissue_class
+@dataclass
+class EvalParams:
+    data: Data
+    device: str
+    pretrained: str
+    model_type: SupervisedModelsArg
+    batch_size: int
+    organ: Organ
 
 
-def random_filter(hdf5_data, percent_to_remove, tissues=None):
-    hdf5_data, mask = hdf5_data.filter_randomly(percent_to_remove)
-    # Remove points from tissue ground truth as well
-    if tissues is not None:
-        tissues = tissues[mask]
-    print(f"Randomly removed {len(tissues) - sum(mask)} points")
-    return hdf5_data, tissues
+@dataclass
+class EvalResult:
+    out: torch.Tensor
+    embeddings: torch.Tensor
 
 
-def confidence_filter(hdf5_data, min_conf, max_conf, tissues=None):
-    hdf5_data, mask = hdf5_data.filter_by_confidence(min_conf, max_conf)
-    # Remove points from tissue ground truth as well
-    if tissues is not None:
-        tissues = tissues[mask]
-    return hdf5_data, tissues
+class EvalRunner:
+    def __init__(self, params: EvalParams):
+        self.params: EvalParams = params
+        self._model: Optional[nn.Module] = None
+        self._loader: Optional[DataLoader] = None
 
-def standardise_cells(hdf5_data):
-    # Standardise cell parts of the data
-    hdf5_data = hdf5_data.standardise_cell_features()
-    return hdf5_data
+        if isinstance(self.model, SIGN_MLP):
+            self.params.data = SIGN(self.model.num_layers)(self.params.data)
 
+    @staticmethod
+    def new(params: EvalParams) -> "EvalRunner":
+        cls = {
+            SupervisedModelsArg.sup_graphsage: LoaderRunner,
+            SupervisedModelsArg.sup_clustergcn: SamplerRunner,
+            SupervisedModelsArg.sup_jumping: SamplerRunner,
+            SupervisedModelsArg.sup_clustergcne: SamplerEdgeRunner,
+            SupervisedModelsArg.sup_clustergin: SamplerEdgeRunner,
+            SupervisedModelsArg.sup_clustergine: SamplerEdgeRunner,
+            SupervisedModelsArg.sup_gat: SamplerRunner,
+            SupervisedModelsArg.sup_gatv2: SamplerRunner,
+            SupervisedModelsArg.sup_graphsaint: SamplerRunner,
+            SupervisedModelsArg.sup_shadow: ShaDowRunner,
+            SupervisedModelsArg.sup_sign: SIGNRunner,
+            SupervisedModelsArg.sup_graphunet: GraphUNetRunner,
+            SupervisedModelsArg.sup_graphunet_noskip: GraphUNetRunner,
+            SupervisedModelsArg.sup_mlp: MLPRunner,
+            SupervisedModelsArg.sup_clustergcn_multilabel: ClusterGCNMultilabelRunner,
+            SupervisedModelsArg.sup_clustergcn_multihead: ClusterGCNMultilabelRunner,
+        }
+        ModelClass = cls[params.model_type]
+        return ModelClass(params)
 
-def get_hdf5_data(
-    project_name, run_id, x_min, y_min, width, height, tissue=False, verbose=True, custom_path = None
-):
-    '''
-    Get HDF5 data from default or custom path
-    '''
-    embeddings_path = get_embeddings_file(project_name, run_id, tissue, custom_path=custom_path)
-    if verbose:
-        print(f"Getting data from: {embeddings_path}")
-        print(f"Using patch of size: x{x_min}, y{y_min}, w{width}, h{height}")
-    # Get hdf5 datasets contained in specified box/patch of WSI
-    hdf5_data = HDF5Dataset.from_path(embeddings_path)
-    hdf5_data = hdf5_data.filter_by_patch(
-        x_min, y_min, width, height
-    ).sort_by_coordinates()
-    return hdf5_data
+    @property
+    def model(self):
+        if self._model is None:
+            _patch_module_getattr()
+            self._model = torch.load(
+                self.params.pretrained, map_location=self.params.device
+            )
+        return self._model
 
+    @property
+    def loader(self):
+        if self._loader is None:
+            self._loader = self.setup_dataloader()
+        return self._loader
 
-def get_groundtruth_patch(organ, project_dir, x_min, y_min, width, height, annot_tsv):
-    if not annot_tsv:
-        print("No tissue annotation tsv supplied")
-        return None, None, None
-    tissue_label_path = project_dir / "results" / "tissue_annots" / annot_tsv
-    if not os.path.exists(str(tissue_label_path)):
-        print("No tissue label tsv found")
-        return None, None, None
-
-    ground_truth_df = pd.read_csv(tissue_label_path, sep="\t")
-    xs = ground_truth_df["px"].to_numpy()
-    ys = ground_truth_df["py"].to_numpy()
-    tissue_classes = ground_truth_df["class"].to_numpy()
-
-    if x_min == 0 and y_min == 0 and width == -1 and height == -1:
-        sort_args = np.lexsort((ys, xs))
-        tissue_ids = np.array(
-            [organ.tissue_by_label(tissue_name).id for tissue_name in tissue_classes]
+    @classmethod
+    def setup_dataloader(cls):
+        raise NotImplementedError(
+            f"setup_dataloader not implemented for {cls.__name__}"
         )
-        return xs[sort_args], ys[sort_args], tissue_ids[sort_args]
 
-    mask = np.logical_and(
-        (np.logical_and(xs > x_min, (ys > y_min))),
-        (np.logical_and(xs < (x_min + width), (ys < (y_min + height)))),
-    )
-    patch_xs = xs[mask]
-    patch_ys = ys[mask]
-    patch_tissue_classes = tissue_classes[mask]
+    @torch.no_grad()
+    def inference(self):
+        self.model.eval()
+        result = self.model_inference()
+        out = result.out
+        predicted_labels = out.argmax(dim=-1, keepdim=True).squeeze()
+        predicted_labels = predicted_labels.cpu().numpy()
+        out = out.cpu().detach().numpy()
+        return out, result.embeddings, predicted_labels
 
-    patch_tissue_ids = np.array(
-        [organ.tissue_by_label(tissue_name).id for tissue_name in patch_tissue_classes]
-    )
-    sort_args = np.lexsort((patch_ys, patch_xs))
+    @classmethod
+    def model_inference(cls) -> EvalResult:
+        """Runs inference across all data in loader."""
+        raise NotImplementedError(f"model_inference not implemented for {cls.__name__}")
 
-    return patch_xs[sort_args], patch_ys[sort_args], patch_tissue_ids[sort_args]
+
+class LoaderRunner(EvalRunner):
+    def setup_dataloader(self):
+        eval_loader = NeighborLoader(
+            self.params.data,
+            num_neighbors=[-1],
+            batch_size=self.params.batch_size,
+            shuffle=False,
+        )
+        eval_loader.data.num_nodes = self.params.data.num_nodes
+        eval_loader.data.n_id = torch.arange(self.params.data.num_nodes)
+        return eval_loader
+
+    def model_inference(self) -> EvalResult:
+        out, embeddings = self.model.inference(
+            self.params.data.x, self.loader, self.params.device
+        )
+        return EvalResult(out=out, embeddings=embeddings)
+
+
+class SamplerRunner(EvalRunner):
+    def setup_dataloader(self):
+        loader = NeighborLoader(
+            self.params.data,
+            num_neighbors=[-1],
+            batch_size=self.params.batch_size,
+            shuffle=False,
+        )
+        loader.data.num_nodes = self.params.data.num_nodes
+        loader.data.n_id = torch.arange(self.params.data.num_nodes)
+        return loader
+
+    def model_inference(self) -> EvalResult:
+        if isinstance(self.model, GraphSAINT):
+            self.model.set_aggr("mean")
+        out, embeddings = self.model.inference(
+            self.params.data.x, self.loader, self.params.device
+        )
+        return EvalResult(out=out, embeddings=embeddings)
+
+
+class SamplerEdgeRunner(SamplerRunner):
+    def model_inference(self) -> EvalResult:
+        if isinstance(self.model, GraphSAINT):
+            self.model.set_aggr("mean")
+        out, embeddings = self.model.inference(
+            self.params.data.x,
+            self.loader,
+            self.params.device,
+        )
+        return EvalResult(out=out, embeddings=embeddings)
+
+
+class ShaDowRunner(EvalRunner):
+    def setup_dataloader(self):
+        return ShaDowKHopSampler(
+            self.params.data,
+            depth=6,
+            num_neighbors=12,
+            node_idx=None,
+            batch_size=self.params.batch_size,
+            shuffle=False,
+        )
+
+    def model_inference(self) -> EvalResult:
+        out = []
+        embeddings = []
+        for batch in self.loader:
+            batch = batch.to(self.params.device)
+            batch_out, batch_embed = self.model.inference(
+                batch.x, batch.edge_index, batch.batch, batch.root_n_id
+            )
+            out.append(batch_out)
+            embeddings.append(batch_embed)
+        out = torch.cat(out, dim=0)
+        embeddings = torch.cat(embeddings, dim=0)
+        return EvalResult(out=out, embeddings=embeddings)
+
+
+class SIGNRunner(EvalRunner):
+    def setup_dataloader(self):
+        return DataLoader(
+            range(self.params.data.num_nodes),
+            batch_size=self.params.batch_size,
+            shuffle=False,
+        )
+
+    def model_inference(self):
+        out = []
+        embeddings = []
+        for idx in self.loader:
+            eval_x = [self.params.data.x[idx].to(self.params.device)]
+            eval_x += [
+                self.params.data[f"x{i}"][idx].to(self.params.device)
+                for i in range(1, self.model.num_layers + 1)
+            ]
+            out_i, emb_i = self.model.inference(eval_x)
+            out.append(out_i)
+            embeddings.append(emb_i)
+        out = torch.cat(out, dim=0)
+        embeddings = torch.cat(embeddings, dim=0)
+        return EvalResult(out=out, embeddings=embeddings)
+
+
+class GraphUNetRunner(EvalRunner):
+    def model_inference(self):
+        out = self.model.inference(
+            self.params.data.x.to(self.params.device),
+            self.params.data.edge_index.to(self.params.device))
+        print("In Graph U-Net embeddings are same as outputs")
+        return EvalResult(out=out, embeddings=out.detach().cpu().clone())
+
+
+class MLPRunner(EvalRunner):
+    def setup_dataloader(self):
+        return DataLoader(
+            range(self.params.data.num_nodes),
+            batch_size=self.params.batch_size,
+            shuffle=False,
+        )
+
+    def model_inference(self):
+        out = []
+        embeddings = []
+        for idx in self.loader:
+            eval_x = self.params.data.x[idx].to(self.params.device)
+            out_i, emb_i = self.model.inference(eval_x)
+            out.append(out_i)
+            embeddings.append(emb_i)
+        out = torch.cat(out, dim=0)
+        embeddings = torch.cat(embeddings, dim=0)
+        return EvalResult(out=out, embeddings=embeddings)
+    
+class ClusterGCNMultilabelRunner(SamplerRunner):
+    @torch.no_grad()
+    def inference(self):
+        self.model.eval()
+        result = self.model_inference()
+        out = result.out
+        out_1 = out[:, :-2]
+        out_2 = out[:, -2:]
+        preds_1 = out_1.argmax(dim=-1)
+        preds_2 = out_2.argmax(dim=-1)
+        preds_2 = preds_2 * -9 + 9
+        preds_1 = preds_1.cpu().numpy()
+        preds_2 = preds_2.cpu().numpy()
+        out_1 = out_1.cpu().detach().numpy()
+        out_2 = out_2.cpu().detach().numpy()
+
+        return out_1, preds_1, out_2, preds_2, result.embeddings
